pyproject.toml
[tool.poetry]
name = "fisheye2equi"
version = "0.1.0"
description = ""
authors = ["ricklon <rianders@docs.rutgers.edu>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
streamlit = "^1.34.0"
opencv-python = "^4.9.0.80"
matplotlib = "^3.8.4"
numpy = "^1.21.5"
Pillow = "^9.0.1"
exif = "^1.3.5"
piexif = "^1.1.3"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
